+++ data/query/draft/keep_latest_pick.sql
