+++ data/query/games/add_series_info.sql
