+++ data/query/playoffs/add_teams_info.sql
