+++ data/query/standings/fix_seeds.sql
